cross_platform_data_transfer_utils/requirements.txt
clickhouse_driver==0.1.3
pulsar_client==2.8.0
kudu_python==1.2.0
fastavro==1.4.5